requirements.txt
Fabric==1.6.1
Django==1.5.1
South==0.8
django-nose==1.1
django-annoying==0.7.7
django_compressor==1.3
django-autocomplete-light==1.2.0
